requirements.txt
flask==3.1.2
gunicorn==23.0.0
gspread==6.1.4
google-auth==2.35.0
